PGDUMP: add a 'HSTORE_COLUMNS' layer creation option (https://trac.osgeo.org/gdal/ticket/4788)
git-svn-id: https://svn.osgeo.org/gdal/trunk@25364 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/pgdump/ogrpgdumpdatasource.cpp
@@ -319,6 +319,9 @@ OGRPGDumpDataSource::CreateLayer( const char * pszLayerName,
 /*      adding tot the srs table if needed.                             */
 /* -------------------------------------------------------------------- */
     int nSRSId = -1;
+    const char* pszPostgisVersion = CSLFetchNameValue( papszOptions, "POSTGIS_VERSION" );
+    if( pszPostgisVersion != NULL && atoi(pszPostgisVersion) >= 2 )
+        nSRSId = 0;
 
     if( CSLFetchNameValue( papszOptions, "SRID") != NULL )
         nSRSId = atoi(CSLFetchNameValue( papszOptions, "SRID"));
@@ -354,15 +357,19 @@ OGRPGDumpDataSource::CreateLayer( const char * pszLayerName,
         else
             pszGFldName = "wkb_geometry";
 
-        /* Sometimes there is an old cruft entry in the geometry_columns
-        * table if things were not properly cleaned up before.  We make
-        * an effort to clean out such cruft.
-        */
-        osCommand.Printf(
-                "DELETE FROM geometry_columns WHERE f_table_name = %s AND f_table_schema = '%s'",
-                pszEscapedTableNameSingleQuote, pszSchemaName );
-        if (bCreateTable)
-            Log(osCommand);
+        if( pszPostgisVersion == NULL || atoi(pszPostgisVersion) < 2 )
+        {
+            /* Sometimes there is an old cruft entry in the geometry_columns
+            * table if things were not properly cleaned up before.  We make
+            * an effort to clean out such cruft.
+            * Note: PostGIS 2.0 defines geometry_columns as a view (no clean up is needed)
+            */
+            osCommand.Printf(
+                    "DELETE FROM geometry_columns WHERE f_table_name = %s AND f_table_schema = '%s'",
+                    pszEscapedTableNameSingleQuote, pszSchemaName );
+            if (bCreateTable)
+                Log(osCommand);
+        }
     }
 
 
@@ -472,6 +479,9 @@ OGRPGDumpDataSource::CreateLayer( const char * pszLayerName,
                                   pszFIDColumnName, nDimension, nSRSId, bWriteAsHex, bCreateTable );
     poLayer->SetLaunderFlag( CSLFetchBoolean(papszOptions,"LAUNDER",TRUE) );
     poLayer->SetPrecisionFlag( CSLFetchBoolean(papszOptions,"PRECISION",TRUE));
+    
+    const char* pszHSTOREColumns = CSLFetchNameValue( papszOptions, "HSTORE_COLUMNS" );
+    poLayer->SetHSTOREColumns(pszHSTOREColumns);
 
 /* -------------------------------------------------------------------- */
 /*      Add layer to data source layer list.                            */
